// app/(tabs)/acara/type.ts
export interface Acara {
  id: string;
  harga: string;
  is_free: boolean;
  pembicara: string;
  is_complete: boolean;
  category: string;
  description: string;
  tempat: string;
  slot: string;
  thumbnailUrl: string;
  name: string;
  tanggal: string;
  jabatan_pembicara: string;
  userId: string;
  is_public: boolean;
}